% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_eps.R
\name{compute_eps}
\alias{compute_eps}
\title{computes the qth quartile of the distribution of distances about the Observed Z Score, Zo, at the true causal model.}
\usage{
compute_eps(Zo, LD, q, weights)
}
\arguments{
\item{Zo}{The Observed Z Score at the SNPs we shall analyse}

\item{LD}{The LD matrix}

\item{q}{The quartile we wish to compute epsilon at}

\item{weights}{The weights we get from LDAK for each SNP}
}
\value{
The value of espilon, the ABC threshold
}
\author{
Mary Fortune
}
