% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{coverage}
\alias{coverage}
\title{coverage of abcGWAS run}
\usage{
coverage(x)
}
\arguments{
\item{x}{object of class ABC}
}
\value{
no return value
}
\description{
print coverage summary, where coverage is defined by the number of times a model was visted
}
\author{
Chris Wallace
}
