% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{credset}
\alias{credset}
\title{Create credible sets}
\usage{
credset(x, size = 0.99, ithr = 2)
}
\arguments{
\item{x}{object of class ABC}

\item{...}{other arguments passed to methods, specifically size
which sets the desired credible set size}
}
\value{
data.table containing snps in the credible set and their
    individual posterior probabilities (po)
}
\description{
Generate 99% credible set of models
}
\details{
.. content for \details{} ..
}
\author{
Chris Wallace
}
