% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{passedfile}
\alias{passedfile}
\title{Internal function}
\usage{
passedfile(d, thr)
}
\arguments{
\item{d}{directory}

\item{thr}{max dist to store}
}
\value{
filename
}
\description{
Functions to return filenames for particular types of file
}
\details{
.. content for \details{} ..
}
\author{
Chris Wallace
}
