% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{ldak.weights}
\alias{ldak.weights}
\title{Calculate SNP weights from a SnpMatrix object}
\usage{
ldak.weights(XX)
}
\arguments{
\item{XX}{SnpMatrix}
}
\description{
This function calls the external executable ldak, which is assumed
to exist in your PATH environment variable.
}
\details{
LDAK can be downloaded from http://dougspeed.com/ldak/.

We used version 4.9.  Later versions may not work if the names of
core arguments changed, but previous versions appear to be kept on
the ldak website.
}
\author{
Chris Wallace, Mary Fortune
}
