% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{plot.ABC}
\alias{plot.ABC}
\title{plotting}
\usage{
plot.ABC(x, what = c("coverage", "nsnp", "ncv", "model"))
}
\arguments{
\item{x}{object of class ABC}

\item{what}{aspect of abcGWAS run to plot
* coverage call plot.coverage (default)
* snp plot mppi summary by threshold
* ncv plot ncv posterior by threshold
* model plot model posterior by threshold}
}
\value{
the plot
}
\author{
Chris Wallace
}
