% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{plot.ABCpost}
\alias{plot.ABCpost}
\title{plot posterior inference}
\usage{
plot.ABCpost(y, po.lower = 0)
}
\arguments{
\item{y}{object of class postABC}

\item{po.lower}{lower threshold - do not display values with posterior probability or (MPPI) < po.lower}
}
\value{
a plot
}
\description{
Plot posterior summary
}
\author{
Chris Wallace
}
