% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{plot.convergence}
\alias{plot.convergence}
\title{plot.convergence}
\usage{
plot.convergence(x, ithr = 2)
}
\arguments{
\item{x}{object of class ABC}

\item{ithr}{which threshold to use. Optional.  Default is second element in thr.}
}
\value{
plot
}
\description{
diagnostic plot to decide whether posterior estimates have converged
}
\author{
Chris Wallace
}
