% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{plot.coverage}
\alias{plot.coverage}
\title{coverage of abcGWAS run}
\usage{
plot.coverage(x)
}
\arguments{
\item{x}{object of class ABC}
}
\value{
no return value
}
\description{
plot histogram of coverage
}
\author{
Chris Wallace
}
