% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{post}
\alias{post}
\title{Posterior inference}
\usage{
post(x, by = c("ncv", "model", "snp"))
}
\arguments{
\item{x}{object of class ABC}

\item{by}{choose to make posterior inference on
* ncv the number of causal variants
* model the choice of W (Var1 and Var2)
* snp generate the marginal posterior probability of inclusion for each SNP}
}
\value{
object of class ABCpost
}
\description{
summarise posterior inference of standard categories
}
\author{
Chris Wallace
}
