% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{preshrink}
\alias{preshrink}
\title{preshrink.}
\usage{
preshrink(d, newsnps = NULL)
}
\arguments{
\item{d}{directory containing abcGWAS output}

\item{newsnps}{don't use (here for test purposes only, will be
deleted)}
}
\value{
0 if success
}
\description{
Given a directory of abcGWAS output, optionally shrink the .RData files
}
\details{
Shrunk files will retain only a subset of the information, and
will be copied to a sub-directory, .shrink, of d.  Doing this once
will make subsequent reading of data faster.
}
\author{
Chris Wallace
}
