% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{run1CV}
\alias{run1CV}
\title{generate ABC samples for single causal variant models}
\usage{
run1CV(Zo, snp.data, weights, freq, eps, outdir, nsam_model = 100,
  nsam_gamma = 1000)
}
\arguments{
\item{Zo}{observed Z score}

\item{snp.data}{SnpMatrix of reference data}

\item{weights}{LDAK weights}

\item{freq}{phased reference haplotypes with Probability column}

\item{eps}{maximum difference between simulated and observed data
for samples to pass}

\item{outdir}{directory within which output file will be created}

\item{nsam_model}{number of causal models to consider}

\item{nsam_gamma}{number of gamma to consider per model}
}
\value{
filename within which samples that passed the threshold
    are written
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Chris Wallace, Mary Fortune
}
