% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{show.summary}
\alias{show.summary}
\title{print a summary of model coverage to date}
\usage{
show.summary(d)
}
\arguments{
\item{d}{directory}
}
\value{
list of nt1 and nt2, as loaded from summary file (invisibly returned)
}
\author{
Chris Wallace
}
