% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{summarise.coverage}
\alias{summarise.coverage}
\title{summarise coverage of an ABC run to date}
\usage{
summarise.coverage(d, force = FALSE)
}
\arguments{
\item{d}{directory containing abc samples}

\item{force}{default FALSE, in which case summary.RData will be regenerated only if one of the abc sample files is newer. If TRUE, regerate anyway}
}
\value{
filename of summary file, returned invisibly
}
\author{
Chris Wallace
}
