% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{summarise_coverage}
\alias{summarise_coverage}
\title{summarise_coverage of an ABC run to date}
\usage{
summarise_coverage(d, filename.only = FALSE, force = FALSE)
}
\arguments{
\item{d}{directory containing abc samples}

\item{force}{default FALSE, in which case summary.RData will be regenerated only if one of the abc sample files is newer. If TRUE, regerate anyway}
}
\value{
filename of summary file
}
\author{
Chris Wallace
}
