% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{summary.ABC}
\alias{summary.ABC}
\title{summary}
\usage{
\method{summary}{ABC}(x)
}
\arguments{
\item{x}{object of class ABC}
}
\value{
no return value
}
\description{
print a brief summary of an abcGWAS run
}
\author{
Chris Wallace
}
