% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{Zstar.dist}
\alias{Zstar.dist}
\title{Calculate distribution of wsumsq(|Zstar|,|Zo|) with Zstar sampled about Zo}
\usage{
Zstar.dist(Zo, snp.data, weights, n = 1e+05)
}
\arguments{
\item{Zo}{observed Z score vector}

\item{snp.data}{reference snp data}

\item{weights}{weights computed by ldak.weights}

\item{n}{specify number of samples to generate distance distribution (optional)}
}
\value{
vector of distances
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Chris Wallace
}
