% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{reader}
\alias{reader}
\title{reader}
\usage{
reader(d, ...)
}
\arguments{
\item{d}{directory or list of directories}

\item{...}{optional parameters passed to inner functions.  Most
useful for test runs is setting n, for example setting n=10
will use only the first 10 files in each directory}
}
\value{
a data.table summarising the passed results and coverage
    so far.  This has class c("ABC","data.table") and some
    convenience functions make use of this additional S3 class.
}
\description{
Read abcGWAS output for a (list of) directory(ies) in a defined format
}
\author{
Chris Wallace
}
